requirements.txt
Django==3.0.8
djangorestframework==3.11.0
django-cors-headers==3.4.0
django-extensions==3.0.2
psycopg2-binary==2.8.5
celery==4.4.6
django-storages==1.9.1
pyjwt==1.7.1
dateutils==0.6.8
watchtower==0.8.0
bleach==3.1.5
django-filter==2.3.0
requests==2.24.0
pytz==2020.1
pycodestyle==2.7.0
